requirements.txt
requests
bs4
python-dotenv
tqdm
